V0.0 update: petit changement

+++ easy_save.Lib/Service/FileSaveService.cs
@@ -60,7 +60,7 @@ namespace easy_save.Lib.Service
 
             foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
             {
-                float progress = (fileCount - fileLeft) * 100 / fileCount;
+                long progress = (fileCount - fileLeft) * 100 / fileCount;
                 progression = progress.ToString() + "%";
                 DateTime before = DateTime.Now;
                 state = "Running";
@@ -128,7 +128,7 @@ namespace easy_save.Lib.Service
                 FileInfo destinationFileInfo = new FileInfo(newPath.Replace(sourcePath, destinationPath));
                 if (!destinationFileInfo.Exists || sourceFileInfo.LastWriteTime > destinationFileInfo.LastWriteTime)
                 {
-                    float progress = (fileCount - fileLeft) * 100 / fileCount;
+                    long progress = (fileCount - fileLeft) * 100 / fileCount;
                     progression = progress.ToString() + "%";
                     DateTime before = DateTime.Now;
                     state = "Running";
